census_app/requirements.txt
alembic==1.16.4
altair==5.5.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
branca==0.8.2
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
decorator==5.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
folium==0.20.0
fonttools==4.60.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
more-itertools==10.7.0
narwhals==2.1.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
plotly==6.3.0
premailer==3.10.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
ratelim==0.1.6
referencing==0.36.2
reportlab==4.4.4
requests==2.32.4
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
streamlit==1.48.0
streamlit-aggrid==1.1.7
streamlit-folium==0.25.3
streamlit-javascript==0.1.5
streamlit-js-eval==0.1.7
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
xlsxwriter==3.2.5
xyzservices==2025.4.0
yagmail==0.15.293
